requirements.txt
# Discord
discord.py>=2.3.0

# LLM
openai>=1.0.0

# API clients
arxiv>=2.0.0
requests>=2.31.0

# Cache & Storage
redis>=5.0.0

# PDF generation
reportlab>=4.0.0
Pillow>=10.0.0

# Scheduling
APScheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging & Monitoring
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0
